+++ packages/salesforcedx-apex-replay-debugger/src/adapter/apexReplayDebug.ts
@@ -23,7 +23,7 @@ import {
 } from 'vscode-debugadapter';
 import { DebugProtocol } from 'vscode-debugprotocol';
 import { breakpointUtil, LineBreakpointEventArgs } from '../breakpoints';
-import { ApexExecutionOverlayResultCommandSuccess } from '../commands';
+
 import {
   GET_LINE_BREAKPOINT_INFO_EVENT,
   LINE_BREAKPOINT_INFO_REQUEST
@@ -164,55 +164,6 @@ export class ScopeContainer extends VariableContainer {
   }
 }
 
-export class ApexHeapDump {
-  private readonly heapDumpId: string;
-  private readonly className: string;
-  private readonly namespace: string;
-  private readonly line: number;
-  private overlaySuccessResut:
-    | ApexExecutionOverlayResultCommandSuccess
-    | undefined;
-  public constructor(
-    heapDumpId: string,
-    className: string,
-    namespace: string,
-    line: number
-  ) {
-    this.heapDumpId = heapDumpId;
-    this.className = className;
-    this.namespace = namespace;
-    this.line = line;
-  }
-  public getHeapDumpId(): string {
-    return this.heapDumpId;
-  }
-  public getClassName(): string {
-    return this.className;
-  }
-  public getNamespace(): string {
-    return this.namespace;
-  }
-  public getLine(): number {
-    return this.line;
-  }
-  public getOverlaySuccessResult():
-    | ApexExecutionOverlayResultCommandSuccess
-    | undefined {
-    return this.overlaySuccessResut;
-  }
-  public setOverlaySuccessResult(
-    overlaySuccessResult: ApexExecutionOverlayResultCommandSuccess
-  ): void {
-    this.overlaySuccessResut = overlaySuccessResult;
-  }
-  /* tslint:disable */
-  public toString = (): string => {
-    return `HeapDumpId: ${this.heapDumpId}, ClassName: ${this
-      .className}, Namespace: ${this.namespace}, Line: ${this.line}`;
-  }; /* This semi-colon is the reason for the tslint, formatting keeps adding it */
-  /* tslint:enable */
-}
-
 export class ApexReplayDebug extends LoggingDebugSession {
   public static THREAD_ID = 1;
   protected logContext: LogContext;
@@ -263,17 +214,13 @@ export class ApexReplayDebug extends LoggingDebugSession {
       nls.localize('session_started_text', this.logContext.getLogFileName())
     );
     // If the projectPath isn't set then don't bother with heap dump processing
-    if (this.projectPath) {
-      if (this.logContext.scanLogForHeapDumpLines()) {
-        if (
-          !await this.logContext.fetchOverlayResultsForApexHeapDumps(
-            this.projectPath
-          )
-        ) {
-          this.errorToDebugConsole(
-            nls.localize('heap_dump_error_wrap_up_text')
-          );
-        }
+    if (this.projectPath && this.logContext.scanLogForHeapDumpLines()) {
+      if (
+        !await this.logContext.fetchOverlayResultsForApexHeapDumps(
+          this.projectPath
+        )
+      ) {
+        this.errorToDebugConsole(nls.localize('heap_dump_error_wrap_up_text'));
       }
     }
     response.success = true;
