+++ packages/salesforcedx-apex-replay-debugger/src/core/index.ts
@@ -7,3 +7,4 @@
 
 export { LogContext } from './logContext';
 export { LogContextUtil } from './logContextUtil';
+export { ApexHeapDump } from './heapDump';
