+++ packages/salesforcedx-apex-replay-debugger/src/messages/i18n.ts
@@ -28,5 +28,9 @@ export const messages = {
   heap_dump_error:
     'Problems were encountered while retrieving heap dump information. Message=%s, ErrorCode=%s. Heap Dump Information=%s',
   heap_dump_error_wrap_up_text:
-    'Problems were encountered while retrieving heap dump information. See the messages above this one for details. Detailed heap dump variable information is unavailable for the specified line or lines.'
+    'Problems were encountered while retrieving heap dump information. See the messages above this one for details. Detailed heap dump variable information is unavailable for the specified line or lines.',
+  fetching_heap_dump:
+    'Retrieving heap dump from server. Heap dump information: %s',
+  malformed_log_line:
+    "Encountered a malformed HEAP_DUMP log line, skipping. Log line number=%d, log line='%s'."
 };
